+++ cppad/local/sweep/for_jac.hpp
@@ -176,7 +176,7 @@ void for_jac(
 			case AbsOp:
 			CPPAD_ASSERT_NARG_NRES(op, 1, 1);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			break;
 			// -------------------------------------------------
@@ -192,7 +192,7 @@ void for_jac(
 			case AddpvOp:
 			CPPAD_ASSERT_NARG_NRES(op, 2, 1);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[1], var_sparsity
+				i_var, size_t(arg[1]), var_sparsity
 			);
 			break;
 			// -------------------------------------------------
@@ -201,7 +201,7 @@ void for_jac(
 			// sqrt(1 - x * x), acos(x)
 			CPPAD_ASSERT_NARG_NRES(op, 1, 2);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			break;
 			// -------------------------------------------------
@@ -211,7 +211,7 @@ void for_jac(
 			// sqrt(x * x - 1), acosh(x)
 			CPPAD_ASSERT_NARG_NRES(op, 1, 2);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			break;
 # endif
@@ -221,7 +221,7 @@ void for_jac(
 			// sqrt(1 - x * x), asin(x)
 			CPPAD_ASSERT_NARG_NRES(op, 1, 2);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			break;
 			// -------------------------------------------------
@@ -231,7 +231,7 @@ void for_jac(
 			// sqrt(1 + x * x), asinh(x)
 			CPPAD_ASSERT_NARG_NRES(op, 1, 2);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			break;
 # endif
@@ -241,7 +241,7 @@ void for_jac(
 			// 1 + x * x, atan(x)
 			CPPAD_ASSERT_NARG_NRES(op, 1, 2);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			break;
 			// -------------------------------------------------
@@ -251,7 +251,7 @@ void for_jac(
 			// 1 - x * x, atanh(x)
 			CPPAD_ASSERT_NARG_NRES(op, 1, 2);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			break;
 # endif
@@ -281,7 +281,7 @@ void for_jac(
 			// sin(x), cos(x)
 			CPPAD_ASSERT_NARG_NRES(op, 1, 2);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			break;
 			// ---------------------------------------------------
@@ -290,7 +290,7 @@ void for_jac(
 			// sinh(x), cosh(x)
 			CPPAD_ASSERT_NARG_NRES(op, 1, 2);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			break;
 			// -------------------------------------------------
@@ -299,7 +299,7 @@ void for_jac(
 			CPPAD_ASSERT_NARG_NRES(op, 2, 1);
 			// derivative is identically zero but dependency is not
 			if( dependency ) forward_sparse_jacobian_unary_op(
-				i_var, arg[1], var_sparsity
+				i_var, size_t(arg[1]), var_sparsity
 			);
 			else
 				var_sparsity.clear(i_var);
@@ -317,7 +317,7 @@ void for_jac(
 			case DivpvOp:
 			CPPAD_ASSERT_NARG_NRES(op, 2, 1);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[1], var_sparsity
+				i_var, size_t(arg[1]), var_sparsity
 			);
 			break;
 			// -------------------------------------------------
@@ -325,7 +325,7 @@ void for_jac(
 			case DivvpOp:
 			CPPAD_ASSERT_NARG_NRES(op, 2, 1);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			break;
 			// -------------------------------------------------
@@ -341,7 +341,7 @@ void for_jac(
 			// arg[0] is always the parameter 2 / sqrt(pi)
 			CPPAD_ASSERT_NARG_NRES(op, 3, 5);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			break;
 			// -------------------------------------------------
@@ -349,7 +349,7 @@ void for_jac(
 			case ExpOp:
 			CPPAD_ASSERT_NARG_NRES(op, 1, 1);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			break;
 			// -------------------------------------------------
@@ -358,7 +358,7 @@ void for_jac(
 			case Expm1Op:
 			CPPAD_ASSERT_NARG_NRES(op, 1, 1);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			break;
 # endif
@@ -419,7 +419,7 @@ void for_jac(
 			case LogOp:
 			CPPAD_ASSERT_NARG_NRES(op, 1, 1);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			break;
 			// -------------------------------------------------
@@ -428,7 +428,7 @@ void for_jac(
 			case Log1pOp:
 			CPPAD_ASSERT_NARG_NRES(op, 1, 1);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			break;
 # endif
@@ -437,7 +437,7 @@ void for_jac(
 			case MulpvOp:
 			CPPAD_ASSERT_NARG_NRES(op, 2, 1);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[1], var_sparsity
+				i_var, size_t(arg[1]), var_sparsity
 			);
 			break;
 			// -------------------------------------------------
@@ -459,7 +459,7 @@ void for_jac(
 			case PowvpOp:
 			CPPAD_ASSERT_NARG_NRES(op, 2, 3);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			break;
 			// -------------------------------------------------
@@ -467,7 +467,7 @@ void for_jac(
 			case PowpvOp:
 			CPPAD_ASSERT_NARG_NRES(op, 2, 3);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[1], var_sparsity
+				i_var, size_t(arg[1]), var_sparsity
 			);
 			break;
 			// -------------------------------------------------
@@ -489,7 +489,7 @@ void for_jac(
 			CPPAD_ASSERT_NARG_NRES(op, 1, 1);
 			// derivative is identically zero but dependency is not
 			if( dependency ) forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			else
 				var_sparsity.clear(i_var);
@@ -500,7 +500,7 @@ void for_jac(
 			// cos(x), sin(x)
 			CPPAD_ASSERT_NARG_NRES(op, 1, 2);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			break;
 			// -------------------------------------------------
@@ -509,7 +509,7 @@ void for_jac(
 			// cosh(x), sinh(x)
 			CPPAD_ASSERT_NARG_NRES(op, 1, 2);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			break;
 			// -------------------------------------------------
@@ -517,7 +517,7 @@ void for_jac(
 			case SqrtOp:
 			CPPAD_ASSERT_NARG_NRES(op, 1, 1);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			break;
 			// -------------------------------------------------
@@ -580,7 +580,7 @@ void for_jac(
 			case SubpvOp:
 			CPPAD_ASSERT_NARG_NRES(op, 2, 1);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[1], var_sparsity
+				i_var, size_t(arg[1]), var_sparsity
 			);
 			break;
 			// -------------------------------------------------
@@ -588,7 +588,7 @@ void for_jac(
 			case SubvpOp:
 			CPPAD_ASSERT_NARG_NRES(op, 2, 1);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			break;
 			// -------------------------------------------------
@@ -597,7 +597,7 @@ void for_jac(
 			// tan(x)^2, tan(x)
 			CPPAD_ASSERT_NARG_NRES(op, 1, 2);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			break;
 			// -------------------------------------------------
@@ -606,7 +606,7 @@ void for_jac(
 			// tanh(x)^2, tanh(x)
 			CPPAD_ASSERT_NARG_NRES(op, 1, 2);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			break;
 			// -------------------------------------------------
@@ -665,7 +665,7 @@ void for_jac(
 			//
 			// argument variables not avaiable during sparsity calculations
 			user_x[user_j]  = CppAD::numeric_limits<Base>::quiet_NaN();
-			user_ix[user_j] = arg[0]; // variable for this argument
+			user_ix[user_j] = size_t(arg[0]); // variable for this argument
 			//
 			++user_j;
 			if( user_j == user_n )
@@ -708,7 +708,7 @@ void for_jac(
 			case ZmulpvOp:
 			CPPAD_ASSERT_NARG_NRES(op, 2, 1);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[1], var_sparsity
+				i_var, size_t(arg[1]), var_sparsity
 			);
 			break;
 			// -------------------------------------------------
@@ -716,7 +716,7 @@ void for_jac(
 			case ZmulvpOp:
 			CPPAD_ASSERT_NARG_NRES(op, 2, 1);
 			forward_sparse_jacobian_unary_op(
-				i_var, arg[0], var_sparsity
+				i_var, size_t(arg[0]), var_sparsity
 			);
 			break;
 			// -------------------------------------------------
